refactor: separate icon module into two
